requirements.txt
google-genai>=0.3.0
vllm>=0.5.4
transformers>=4.43.0
accelerate>=0.33.0
torch>=2.2.0; platform_system != 'Windows'
